backend/requirements.txt
# Database & Security
SQLAlchemy>=2.0.0
pyodbc>=5.0.0
authlib>=1.3.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0

# Azure Stack
azure-storage-blob>=12.19.0
azure-cognitiveservices-speech>=1.34.0

# AI Services
deepgram-sdk>=3.0.0
google-generativeai>=0.4.0
openai>=1.12.0

# Audio Processing
pydub>=0.25.1
aiofiles>=23.2.1

# HTTP & Async
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.8.0
pydantic-settings>=2.3.0
python-dotenv==1.0.0
aiohttp>=3.8.0
httpx>=0.26.0
python-multipart>=0.0.9

# Utilities
python-json-logger>=2.0.0
numpy>=2.0.0
scipy>=1.13.0
